Refactor liveblog specific hydration to liveblog.ts
